[accumulo] move accumulo into a dedicated package.
